LayoutTests/inspector/editor/highlighter-paste-in-comment-expected.txt

This test checks against a bug in the highlighter when pasting text inside a comment.


After highlighting all lines
1 : /* : javascript-comment[0-3]
2 :  :
3 : 1 : javascript-comment[0-1]
4 : 2 : javascript-comment[0-1]
5 : 3 : javascript-comment[0-1]
6 : 4 : javascript-comment[0-1]
7 : 5 : javascript-comment[0-1]
8 :  :
9 : */ : javascript-comment[0-2]

After update highlights
1 : /* : javascript-comment[0-3]
2 :  :
3 : 1 : javascript-comment[0-1]
4 : 2 : javascript-comment[0-1]
5 : 3 : javascript-comment[0-1]
6 : 4 : javascript-comment[0-1]
7 : 5 : javascript-comment[0-1]
8 : 1 : javascript-comment[0-1]
9 : 2 : javascript-comment[0-1]
10 : 3 : javascript-comment[0-1]
11 : 4 : javascript-comment[0-1]
12 : 5 : javascript-comment[0-1]
13 :  :
14 : */ : javascript-comment[0-2]

After highlighting up to the end
1 : /* : javascript-comment[0-3]
2 :  :
3 : 1 : javascript-comment[0-1]
4 : 2 : javascript-comment[0-1]
5 : 3 : javascript-comment[0-1]
6 : 4 : javascript-comment[0-1]
7 : 5 : javascript-comment[0-1]
8 : 1 : javascript-comment[0-1]
9 : 2 : javascript-comment[0-1]
10 : 3 : javascript-comment[0-1]
11 : 4 : javascript-comment[0-1]
12 : 5 : javascript-comment[0-1]
13 :  :
14 : */ : javascript-comment[0-2]

After inserting a new line
1 : /* : javascript-comment[0-3]
2 :  :
3 : 1 : javascript-comment[0-1]
4 : 2 : javascript-comment[0-1]
5 : 3 : javascript-comment[0-1]
6 : 4 : javascript-comment[0-1]
7 : 5 : javascript-comment[0-1]
8 : 1 : javascript-comment[0-1]
9 : 2 : javascript-comment[0-1]
10 : 3 : javascript-comment[0-1]
11 : 4 : javascript-comment[0-1]
12 : 5 : javascript-comment[0-1]
13 :  :
14 :  :
15 : */ : javascript-comment[0-2]
